mobile: initial commit (empty quasar project)
Nothing added on top of the standard instructions for building this.

How to get here:
1. Setup your favorite node-version-manager (nvm in this case)
    a. go to: https://github.com/nvm-sh/nvm#installing-and-updating
    b. [at this point in time] run
        curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.36.0/install.sh | bash
    c. finish setup as the command tells you (shell setup, bash completion, etc)
2. run: nvm install --lts
3. run: nvm use --lts
4. Go to: https://quasar.dev/quasar-cli/installation
    [yep, using Quasar CLI for this project]
    a. run:  npm install -g @quasar/cli
5. run: quasar create mobile
6. Select some default settings
7. run: cd mobile
8. run: quasar dev

Steps 1-to-3 are needed for a first-time NVM/Nodejs setup.
Steps 4 is needed for the first-time Quasar setup.
Steps 5-to-6 are needed for the first-time project setup.
Steps 7-to-8 are needed everytime the project is run.

On top of these steps, 'git init' was run to create this first commit.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>